pyproject.toml
[project]
name = "graph-rag"
version = "0.1.0"
description = "MSCI SEC filing chunking, Neo4j + Azure embeddings, and Form 13F investor graph"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.2.1",
    "neo4j>=6.1.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
